Make integer settings use IntegerListPreference
Avoids unnecessary type casting of setting values and duplication in resource files.

+++ app/src/main/java/emu/skyline/SkylineApplication.kt
@@ -36,6 +36,6 @@ class SkylineApplication : Application() {
 
         val appFilesPath = applicationContext.getPublicFilesDir().canonicalPath
         File("$appFilesPath/logs/").mkdirs()
-        initializeLog("$appFilesPath/", getSettings().logLevel.toInt())
+        initializeLog("$appFilesPath/", getSettings().logLevel)
     }
 }
